Working on addition.

+++ code/ubigint.cpp
@@ -46,12 +46,12 @@ ubigint ubigint::operator+ (const ubigint& that) const {
         if (i > (that.ubig_value.size() - 1)) {
             digit1 = 0;
         } else {
-            digit1 = static_cast<int>(that.ubig_value[i]) - 48;
+            digit1 = that.ubig_value[i] - 48;
         }
         if (i > (ubig_value.size() - 1)) {
             digit2 = 0;
         } else {
-            digit2 = static_cast<int>(ubig_value[i]) - 48;
+            digit2 = ubig_value[i] - 48;
         }
         partial_result = digit1 + digit2;
         cout << digit1 << " + " << digit2 << " = " << partial_result << endl;
